+++ src/xyz/oihana/schema/constants/traits/ThesaurusTrait.php
@@ -2,6 +2,7 @@
 
 namespace xyz\oihana\schema\constants\traits;
 
+use xyz\oihana\schema\constants\traits\thesaurus\CollectionTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ConceptSchemeTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ConceptTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\SkosMappingsTrait;
@@ -17,7 +18,8 @@ use xyz\oihana\schema\constants\traits\thesaurus\ThesaurusTermTrait;
  */
 trait ThesaurusTrait
 {
-    use ConceptSchemeTrait ,
+    use CollectionTrait ,
+        ConceptSchemeTrait ,
         ConceptTrait ,
         SkosMappingsTrait ,
         SkosNotesTrait ,